requirements.txt
#requirements.txt

python-dotenv
Flask
pandas
requests
